// types.ts
export interface SignatureData {
  name: string;
  role: string;
  email: string;
  phoneFixe: string;
  mobile: string;
  website: string;
  address: string;
  logoType: 'svg' | 'url';
  logoContent: string;
  width: number;
  nameFontSize: number;
  roleFontSize: number;
  contactFontSize: number;
}